requirements.txt
Jinja2
shyaml
PyYAML~=5.3.1
file-downloader
libvirt-python
cm-client